Start clean up code
