+++ pykib.py
@@ -1,6 +1,8 @@
 # pykib - A PyQt5 based kiosk browser with a minimum set of functionality
 # Copyright (C) 2018 Tobias Wintrich
 #
+# This file is part of pykib.
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, either version 3 of the License, or
@@ -20,11 +22,14 @@ import os
 import textwrap
 from functools import partial
 from PyQt5 import QtCore, QtWidgets, QtWebEngineWidgets
-from PyQt5.QtCore import QSize, QUrl, QFile
+from PyQt5.QtCore import QSize, QUrl
 from PyQt5.QtGui import QIcon, QKeyEvent
-from PyQt5.QtWebEngineWidgets import QWebEngineView, QWebEnginePage, QWebEngineProfile
+#from PyQt5.QtWebEngineWidgets import QWebEngineView, QWebEnginePage, QWebEngineProfile
 from PyQt5.QtWidgets import QApplication, QWidget 
 from argparse import ArgumentParser, RawDescriptionHelpFormatter
+
+from pykib_base.myQWebEngineView import myQWebEngineView
+from pykib_base.myQWebEnginePage import myQWebEnginePage
 
 __version_info__ = ('beta', '0.9')
 __version__ = '-'.join(__version_info__)
@@ -66,10 +71,10 @@ class MainWindow(QWidget):
         self.navGridLayout.setContentsMargins(9, 9, 9, 0)
         self.navGridLayout.setObjectName("navGridLayout")        
         
-        self.web = WebView()
+        self.web = myQWebEngineView()
         self.web.setObjectName("view")
         
-        self.page = WebViewPage(self.web)
+        self.page = myQWebEnginePage(self.web)
         
         self.web.setPage(self.page)
         #self.page.setView(self.web)
@@ -154,110 +159,7 @@ class MainWindow(QWidget):
         
     def adjustAdressbar(self):
         self.addressBar.setText(self.web.url().toString())
-    
-
-        
-   
-class WebView(QWebEngineView):
-    def __init__(self):
-        self.browser = QWebEngineView.__init__(self)        
-        self.setContextMenuPolicy( QtCore.Qt.NoContextMenu )
-        
-    def load(self,url):
-        if not url:
-            url = args.url
-        if not (url.startswith('http://') or url.startswith('https://') or url.startswith('file:///')):
-            url = 'http://' + url        
-        self.setUrl(QUrl(url))      
-    
-
-class WebViewPage(QWebEnginePage):
-    def __init__(self, *args, **kwargs):
-        QtWebEngineWidgets.QWebEnginePage.__init__(self, *args, **kwargs)
-        self.profile().downloadRequested.connect(self.on_downloadRequested)
-        #Do not persist Cookies
-        self.profile().setPersistentCookiesPolicy(QWebEngineProfile.NoPersistentCookies)
-    
-    def javaScriptConsoleMessage(self, msg, lineNumber, sourceID, category):
-        #Ignore JS Failures
-        pass
-        
-    #Certificate Error handling
-    def certificateError(self, error):
-        if(args.ingoreCertificates):
-            print("Certificate Error")
-            return True
-        else:
-            return False
-    
-    #Download Handle
-    @QtCore.pyqtSlot(QtWebEngineWidgets.QWebEngineDownloadItem)
-    def on_downloadRequested(self, download):
-        downloadHandleHit = False
-        old_path = download.path()
-        suffix = QtCore.QFileInfo(old_path).suffix()
-        for x in args.downloadHandle:
-            handle = x.split("|")
-            if(suffix == handle[0]):    
-                downloadHandleHit = True
-                filepath = handle[2]+"/tmp."+suffix                
-                download.setPath(filepath)
-                download.accept()
-                download.finished.connect(partial(self.runProcess, handle, filepath, download))                
-                
-        if(args.download and not downloadHandleHit):            
-            path, _ = QtWidgets.QFileDialog.getSaveFileName(self.view(), "Save File", old_path, "*."+suffix)
-            if path:
-                download.setPath(path)
-                download.accept()
-                
-    def runProcess(self, handle, filepath, download):
-        print(download.isFinished())
-        print("Executing:" +handle[1] +" "+filepath);
-        subprocess.Popen(handle[1] +" "+filepath, shell=True)
-    
-    def acceptNavigationRequest(self, url: QUrl, typ: QWebEnginePage.NavigationType, is_main_frame: bool):
-        if(args.whiteList):
-            return self.checkWhitelist(url)
-        else:
-            return True
-        
-    def checkWhitelist(self, url:QUrl): 
-        currentUrl = url.toString()
-        for x in args.whiteList:
-            if(currentUrl.startswith(x)):
-                return True;
-        print("Site "+ currentUrl +" is not whitelisted")       
-        
-        msg = QtWidgets.QMessageBox()
-        msg.setIcon(QtWidgets.QMessageBox.Warning)
-        msg.setText( "Site "+ currentUrl +" is not white-listed")
-        msg.setWindowTitle("Whitelist Error")
-        
-            
-        backButton = QtWidgets.QPushButton("Go Back")
-        backButton.setIcon(QIcon(os.path.join(dirname, 'icons/back.png')));
-        backButton.setIconSize(QSize(24, 24));
-        backButton.setObjectName("backButton")
-        
-        msg.addButton(backButton, QtWidgets.QMessageBox.NoRole )
-        
-        homeButton = QtWidgets.QPushButton("Go Home")
-        homeButton.setIcon(QIcon(os.path.join(dirname, 'icons/home.png')));
-        homeButton.setIconSize(QSize(24, 24));
-        homeButton.setObjectName("homeButton")
-        
-        msg.addButton(homeButton, QtWidgets.QMessageBox.NoRole )
-        
-        msg.show()
-        retval = msg.exec_()
-        
-        if(retval == 0):            
-            self.view().stop()
-        else:
-            self.view().load(args.url)
-        
-        return False
+ 
 
 parser = ArgumentParser(
       prog='pykib',
